+++ tests/KingmakerGunslinger.DomainTests/Program.cs
@@ -77,6 +77,7 @@ namespace KingmakerGunslinger.DomainTests
             Case("elemental-traits.blood-insight-save-matrix", ElementalBloodInsightPersistencePolicyTests.NativeFixtureCoverageIsExact),
             Case("elemental-traits.efreeti-combined-save-matrix", ElementalBloodInsightPersistencePolicyTests.EfreetiCombinedFixtureCoverageIsExact),
             Case("elemental-traits.crystalline-save-matrix", ElementalBloodInsightPersistencePolicyTests.CrystallineFixtureCoverageIsExact),
+            Case("elemental-traits.breath-save-matrix", ElementalBloodInsightPersistencePolicyTests.BreathFixtureCoverageIsExact),
             Case("elemental-traits.daily-resource-reconstruction", ElementalTraitDailyResourcePolicyTests.ActivationNeverRefillsSpentUses),
             Case("elemental-traits.breath-damage-breakpoints", ElementalBreathPolicyTests.DamageBreakpointsAndCapAreExact),
             Case("elemental-traits.breath-dc-current-stat", ElementalBreathPolicyTests.DifficultyUsesCurrentConstitutionAndUncappedHalfLevel),